+++ sync-for-expenses/src/sdk/sdk.ts
@@ -58,8 +58,9 @@ export class SDKConfiguration {
     serverDefaults: any;
     language = "typescript";
     openapiDocVersion = "prealpha";
-    sdkVersion = "2.2.1";
-    genVersion = "2.131.1";
+    sdkVersion = "2.3.0";
+    genVersion = "2.139.1";
+    userAgent = "speakeasy-sdk/typescript 2.3.0 2.139.1 prealpha @codat/sync-for-expenses";
     retryConfig?: utils.RetryConfig;
     public constructor(init?: Partial<SDKConfiguration>) {
         Object.assign(this, init);